Moved set_* packets to packet/settings folder
