+++ server/src/network/packets/settings/mod.rs
@@ -1,0 +1,10 @@
+use common::glob_export;
+
+glob_export!(set_commands_enabled);
+glob_export!(set_default_game_mode);
+glob_export!(set_difficulty);
+glob_export!(set_player_gamemode);
+glob_export!(set_scoreboard_identity);
+glob_export!(set_time);
+glob_export!(set_title);
+glob_export!(update_skin);
